+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## Version 2.1.1
 
 * [FIXED] Issue with version in Nats.java, also updated deploying.md with checklist
+* [FIXED] Fixed issue during reconnect where buffered messages blocked protocol messages
 
 ## Version 2.1.0